Arrange components using constraing layout
Constraint layout is the modern approach to create layouts for your
apps. If you don't use the constrainsts, when you run your app it will
misplace the components. You can align a component relative to its
parent or relative to another component. You can do it by hand or using
the android UI.